+++ state-transition/core/core_test.go
@@ -23,6 +23,7 @@ package core_test
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"testing"
 
 	corestore "cosmossdk.io/core/store"
@@ -32,6 +33,7 @@ import (
 	storetypes "cosmossdk.io/store/types"
 	"github.com/berachain/beacon-kit/chain-spec/chain"
 	"github.com/berachain/beacon-kit/consensus-types/types"
+	engineprimitives "github.com/berachain/beacon-kit/engine-primitives/engine-primitives"
 	"github.com/berachain/beacon-kit/log/noop"
 	"github.com/berachain/beacon-kit/node-core/components"
 	nodemetrics "github.com/berachain/beacon-kit/node-core/components/metrics"
@@ -228,9 +230,66 @@ func buildNextBlock(
 	}
 }
 
-// We verify DepositRoot in Eth1Data, but we don't really update
-// the store once we store deposits (it's not StateProcessor responsibility).
-// So we just keep using the dummyEth1Data most of the time.
-var dummyEth1Data = &types.Eth1Data{
-	DepositRoot: types.Deposits(nil).HashTreeRoot(),
+func generateTestExecutionAddress(
+	t *testing.T,
+	rndSeed int,
+) (types.WithdrawalCredentials, int) {
+	t.Helper()
+
+	addrStr := strconv.Itoa(rndSeed)
+	addrBytes := bytes.ExtendToSize([]byte(addrStr), bytes.B20Size)
+	execAddr, err := bytes.ToBytes20(addrBytes)
+	require.NoError(t, err)
+	rndSeed++
+	return types.NewCredentialsFromExecutionAddress(
+		common.ExecutionAddress(execAddr),
+	), rndSeed
+}
+
+func generateTestPK(t *testing.T, rndSeed int) (bytes.B48, int) {
+	t.Helper()
+	keyStr := strconv.Itoa(rndSeed)
+	keyBytes := bytes.ExtendToSize([]byte(keyStr), bytes.B48Size)
+	key, err := bytes.ToBytes48(keyBytes)
+	require.NoError(t, err)
+	rndSeed++
+	return key, rndSeed
+}
+
+func moveToEndOfEpoch(
+	t *testing.T,
+	tip *types.BeaconBlock,
+	cs chain.Spec[bytes.B4, math.U64, math.U64, any],
+	sp *TestStateProcessorT,
+	st *TestBeaconStateT,
+	ctx *transition.Context,
+	depRoot common.Root,
+) *types.BeaconBlock {
+	t.Helper()
+	blk := tip
+	currEpoch := cs.SlotToEpoch(blk.GetSlot())
+	for currEpoch == cs.SlotToEpoch(blk.GetSlot()+1) {
+		blk = buildNextBlock(
+			t,
+			st,
+			&types.BeaconBlockBody{
+				ExecutionPayload: &types.ExecutionPayload{
+					Timestamp:    blk.Body.ExecutionPayload.Timestamp + 1,
+					ExtraData:    []byte("testing"),
+					Transactions: [][]byte{},
+					Withdrawals: []*engineprimitives.Withdrawal{
+						st.EVMInflationWithdrawal(),
+					},
+					BaseFeePerGas: math.NewU256(0),
+				},
+				Eth1Data: &types.Eth1Data{DepositRoot: depRoot},
+				Deposits: []*types.Deposit{},
+			},
+		)
+
+		vals, err := sp.Transition(ctx, st, blk)
+		require.NoError(t, err)
+		require.Empty(t, vals) // no vals changes expected before next epoch
+	}
+	return blk
 }
