+++ state-transition/core/state_processor_genesis.go
@@ -57,8 +57,6 @@ func (sp *StateProcessor[
 		return nil, err
 	}
 
-	// Eth1DepositIndex will be set in processDeposit
-
 	var eth1Data *ctypes.Eth1Data
 	eth1Data = eth1Data.New(
 		deposits.HashTreeRoot(),
@@ -95,6 +93,10 @@ func (sp *StateProcessor[
 		}
 	}
 
+	// Before processing deposits, set the eth1 deposit index to 0.
+	if err := st.SetEth1DepositIndex(0); err != nil {
+		return nil, err
+	}
 	if err := sp.validateGenesisDeposits(st, deposits); err != nil {
 		return nil, err
 	}
